error handling in getEmotionData api

+++ app.py
@@ -59,15 +59,19 @@ def get_state():
 def get_emotion_data():
     data = json.loads(request.get_data())
     user_id = data["user_id"]
+    datapath = f'emotion_data/{user_id}.json'
 
-    with open(f'emotion_data/{user_id}.json', 'r') as f:
-        _emotion_data = json.load(f)
-    emotion_data = _emotion_data["records"]
-    _emotion_data["records"].clear()
-    with open(f'emotion_data/{user_id}.json', 'w') as f:
-        json.dump(_emotion_data, f)
+    if os.path.exists(datapath):
+        with open(datapath, 'r') as f:
+            _emotion_data = json.load(f)
+        emotion_data = _emotion_data["records"]
+        _emotion_data["records"].clear()
+        with open(datapath, 'w') as f:
+            json.dump(_emotion_data, f)
+        return jsonify({"emotion_data": emotion_data})
 
-    return jsonify({"emotion_data": emotion_data})
+    else:
+        return jsonify({"error": f'No user information: {user_id}'})
 
 if __name__ == "__main__":
     app.run(debug=True, host='0.0.0.0', port=80)
